requirements.txt
ffmpeg-python==0.2.0
pyinstaller==6.10.0